Improve handling of hg data in tests
Merurial uses a weird mixture of line endings which is hard to
get right when patching the commits and working on different
platforms. This commit (and an accompanying commit that cleans up
the test data itself) will change the way we handle the test data:
we only store those files that are not in hg, plus the `.hg`
subdirectory itself. When running the tests, we copy the test data
directory, but now also checkout the tip of the hg repo.